+++ scripts/autotracking/archipelago.lua
@@ -1345,7 +1345,7 @@ function hostedLocationsHighlight(location_code, highlight_code)
         objItem.Highlight = highlight_code
         objItem = Tracker:FindObjectForCode("@Azure Archipelago/Azure Archipelago - Weapon Shop/Cross Goggles Type - 150000 Credits")
         objItem.Highlight = highlight_code
-        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Goggles Type - 150000 Credits")
+        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Goggles Type - 150000 Credits Have Azure A. Pass")
         objItem.Highlight = highlight_code
 
     elseif location_code == "@Shop Types/Cross Edge Type/Buy A 'Cross Edge' From Any Item Shop" then
@@ -1353,7 +1353,7 @@ function hostedLocationsHighlight(location_code, highlight_code)
         objItem.Highlight = highlight_code
         objItem = Tracker:FindObjectForCode("@Azure Archipelago/Azure Archipelago - Weapon Shop/Cross Edge Type - 150000 Credits")
         objItem.Highlight = highlight_code
-        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Edge Type - 150000 Credits")
+        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Edge Type - 150000 Credits Have Azure A. Pass")
         objItem.Highlight = highlight_code
 
     elseif location_code == "@Shop Types/Cross Mail Type/Buy A 'Cross Mail' From Any Item Shop" then
@@ -1361,7 +1361,7 @@ function hostedLocationsHighlight(location_code, highlight_code)
         objItem.Highlight = highlight_code
         objItem = Tracker:FindObjectForCode("@Azure Archipelago/Azure Archipelago - Weapon Shop/Cross Mail Type - 150000 Credits")
         objItem.Highlight = highlight_code
-        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Mail Type - 150000 Credits")
+        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Mail Type - 150000 Credits Have Azure A. Pass")
         objItem.Highlight = highlight_code
 
     elseif location_code == "@Shop Types/Cross Boots Type/Buy A 'Cross Boots' From Any Item Shop" then
@@ -1369,7 +1369,7 @@ function hostedLocationsHighlight(location_code, highlight_code)
         objItem.Highlight = highlight_code
         objItem = Tracker:FindObjectForCode("@Azure Archipelago/Azure Archipelago - Weapon Shop/Cross Boots Type - 150000 Credits")
         objItem.Highlight = highlight_code
-        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Boots Type - 150000 Credits")
+        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Boots Type - 150000 Credits Have Azure A. Pass")
         objItem.Highlight = highlight_code
 
     elseif location_code == "@Shop Types/Goat Cheese Type/Buy A 'Goat Cheese' From Any Item Shop" then
@@ -2498,32 +2498,32 @@ function manualShopTypes(location_id)
         objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Curio Shop/Chest Detector Type - 33333 Credits")
         objItem.AvailableChestCount = objItem.AvailableChestCount - 1
     elseif location_id == 3235824901 then
-        --objItem = Tracker:FindObjectForCode("@Homestedt/Homestedt - Weapon Shop/Cross Goggles Type - 150000 Credits")
-        --objItem.AvailableChestCount = objItem.AvailableChestCount - 1
+        objItem = Tracker:FindObjectForCode("@Homestedt/Homestedt - Weapon Shop/Cross Goggles Type - 150000 Credits Have Azure A. Pass")
+        objItem.AvailableChestCount = objItem.AvailableChestCount - 1
         objItem = Tracker:FindObjectForCode("@Azure Archipelago/Azure Archipelago - Weapon Shop/Cross Goggles Type - 150000 Credits")
         objItem.AvailableChestCount = objItem.AvailableChestCount - 1
-        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Goggles Type - 150000 Credits")
+        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Goggles Type - 150000 Credits Have Azure A. Pass")
         objItem.AvailableChestCount = objItem.AvailableChestCount - 1
     elseif location_id == 3235824903 then
-        --objItem = Tracker:FindObjectForCode("@Homestedt/Homestedt - Weapon Shop/Cross Edge Type - 150000 Credits")
-        --objItem.AvailableChestCount = objItem.AvailableChestCount - 1
+        objItem = Tracker:FindObjectForCode("@Homestedt/Homestedt - Weapon Shop/Cross Edge Type - 150000 Credits Have Azure A. Pass")
+        objItem.AvailableChestCount = objItem.AvailableChestCount - 1
         objItem = Tracker:FindObjectForCode("@Azure Archipelago/Azure Archipelago - Weapon Shop/Cross Edge Type - 150000 Credits")
         objItem.AvailableChestCount = objItem.AvailableChestCount - 1
-        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Edge Type - 150000 Credits")
+        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Edge Type - 150000 Credits Have Azure A. Pass")
         objItem.AvailableChestCount = objItem.AvailableChestCount - 1
     elseif location_id == 3235824905 then
-        --objItem = Tracker:FindObjectForCode("@Homestedt/Homestedt - Weapon Shop/Cross Mail Type - 150000 Credits")
-        --objItem.AvailableChestCount = objItem.AvailableChestCount - 1
+        objItem = Tracker:FindObjectForCode("@Homestedt/Homestedt - Weapon Shop/Cross Mail Type - 150000 Credits Have Azure A. Pass")
+        objItem.AvailableChestCount = objItem.AvailableChestCount - 1
         objItem = Tracker:FindObjectForCode("@Azure Archipelago/Azure Archipelago - Weapon Shop/Cross Mail Type - 150000 Credits")
         objItem.AvailableChestCount = objItem.AvailableChestCount - 1
-        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Mail Type - 150000 Credits")
+        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Mail Type - 150000 Credits Have Azure A. Pass")
         objItem.AvailableChestCount = objItem.AvailableChestCount - 1
     elseif location_id == 3235824907 then
-        --objItem = Tracker:FindObjectForCode("@Homestedt/Homestedt - Weapon Shop/Cross Boots Type - 150000 Credits")
-        --objItem.AvailableChestCount = objItem.AvailableChestCount - 1
+        objItem = Tracker:FindObjectForCode("@Homestedt/Homestedt - Weapon Shop/Cross Boots Type - 150000 Credits Have Azure A. Pass")
+        objItem.AvailableChestCount = objItem.AvailableChestCount - 1
         objItem = Tracker:FindObjectForCode("@Azure Archipelago/Azure Archipelago - Weapon Shop/Cross Boots Type - 150000 Credits")
         objItem.AvailableChestCount = objItem.AvailableChestCount - 1
-        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Boots Type - 150000 Credits")
+        objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Weapon Shop/Cross Boots Type - 150000 Credits Have Azure A. Pass")
         objItem.AvailableChestCount = objItem.AvailableChestCount - 1
     elseif location_id == 3235824913 then
         objItem = Tracker:FindObjectForCode("@Rhombus Square/Rhombus Square - Goat Zira's Shop/Goat Cheese Type - 4500 Credits")